+++ dbms/postgresql/src/main/java/is/codion/dbms/postgresql/PostgreSQLDatabase.java
@@ -103,22 +103,22 @@ final class PostgreSQLDatabase extends AbstractDatabase {
 
   @Override
   public boolean isAuthenticationException(SQLException exception) {
-    return INVALID_PASS.equals(exception.getSQLState());
+    return INVALID_PASS.equals(requireNonNull(exception).getSQLState());
   }
 
   @Override
   public boolean isReferentialIntegrityException(SQLException exception) {
-    return FOREIGN_KEY_VIOLATION.equals(exception.getSQLState());
+    return FOREIGN_KEY_VIOLATION.equals(requireNonNull(exception).getSQLState());
   }
 
   @Override
   public boolean isUniqueConstraintException(SQLException exception) {
-    return UNIQUE_CONSTRAINT_ERROR.equals(exception.getSQLState());
+    return UNIQUE_CONSTRAINT_ERROR.equals(requireNonNull(exception).getSQLState());
   }
 
   @Override
   public boolean isTimeoutException(SQLException exception) {
-    return TIMEOUT_ERROR.equals(exception.getSQLState());
+    return TIMEOUT_ERROR.equals(requireNonNull(exception).getSQLState());
   }
 
   @Override
@@ -133,6 +133,8 @@ final class PostgreSQLDatabase extends AbstractDatabase {
 
   @Override
   public String errorMessage(SQLException exception, Operation operation) {
+    requireNonNull(exception);
+    requireNonNull(operation);
     String sqlState = exception.getSQLState();
     if (NULL_VALUE_ERROR.equals(sqlState)) {
       return createNullValueErrorMessage(exception.getMessage());
@@ -140,8 +142,8 @@ final class PostgreSQLDatabase extends AbstractDatabase {
     if (UNIQUE_CONSTRAINT_ERROR.equals(sqlState)) {
       return createUniqueConstraintErrorMessage(exception.getMessage());
     }
-    if (FOREIGN_KEY_VIOLATION.equals(sqlState) && operation == Operation.DELETE) {
-      return ERROR_CODE_MAP.get(FOREIGN_KEY_VIOLATION_DELETE);
+    if (FOREIGN_KEY_VIOLATION.equals(sqlState)) {
+      return createForeignKeyViolationErrorMessage(operation);
     }
     if (ERROR_CODE_MAP.containsKey(sqlState)) {
       return ERROR_CODE_MAP.get(sqlState);
@@ -169,10 +171,18 @@ final class PostgreSQLDatabase extends AbstractDatabase {
     if (indexOfDetail != -1 && indexOfAlreadyExists != -1) {
       //Detail: Key (col1, col2)=(val1, val2) already exists.
       String values = exceptionMessage.substring(indexOfDetail + 11, indexOfAlreadyExists);
-      
+
       return MESSAGES.getString(UNIQUE_KEY_ERROR) + ": " + values;
     }
 
     return MESSAGES.getString(UNIQUE_KEY_ERROR);
+  }
+
+  private static String createForeignKeyViolationErrorMessage(Operation operation) {
+    if (operation == Operation.DELETE) {
+      return ERROR_CODE_MAP.get(FOREIGN_KEY_VIOLATION_DELETE);
+    }
+
+    return ERROR_CODE_MAP.get(FOREIGN_KEY_VIOLATION);
   }
 }